+++ tests/settings.py
@@ -7,6 +7,19 @@ TEMPLATE_DIRS = [
 
 TEMPLATE_DEBUG = True
 
+# Django >=1.9
+TEMPLATES = [
+    {
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        'DIRS': TEMPLATE_DIRS,
+        'APP_DIRS': True,
+        'OPTIONS': {
+            'debug': TEMPLATE_DEBUG,
+        }
+    }
+]
+
+
 SECRET_KEY = "foobar"
 INSTALLED_APPS = [
     'django.contrib.auth',